v2.0.0: automorphism comparison (#16)
* chore: add enviroment configuration and documentation

* doc: add comments

* feat: add new architecture and initial worker implementation (#1)

* feat: add worker (no result saving)

* docs: add pipeline architecture diagram

* feat: update architecture to use one SQS (#2)

* feat: add orchestrator structure (#3)

* feat: ochestrator structure

* feat: add orchestrator structure

* feat: add script to upload error sample arrays JSON to S3 (#6)

* feat: add architecture development designed on July 1st using Lambda, DynamoDb local and Backblaze. (#10)

* chore: add schema of new architecture and reorganize directories.

* feat: add orchestrator deployed with Backblaze and local DynamoDB

* feat(architecture): Initial implementation of architecture scripts without testing.

* feat: add local deployments without invocations between lambdas

* feat: integrate services and save results in DynamoDB (#11)

* wip: test-with-serverless-offline-plugin

* feat: add integration between args_mixer and orchestrator

* feat: add full pipeline

* feat/replace backblaze with serverless s3 local and complete pipeline (#12)

* fix: seed usability

* feat/add_enviroment_files

* fix: save last batch even if smaller than batch size

* chore: prepare release

* feat: implement a new architecture for scanning multiple automorphisms. (#15)

* wip: scan of different automorphisms

* feat: implement scan over multiple automorphisms

* feat: add documentation for new architecture

* doc: add pending documentation

* doc: add comments

* docs: update local deployment instructions and add experimentation results

* doc: add logs file

* feat: clean repo

* docs: clean docs dir

+++ resources/manage_resources.py
@@ -89,12 +89,20 @@ def clean_s3_bucket(bucket_name):
     s3_client = get_s3_client()
     
     try:
-        # Listar y eliminar todos los objetos en el bucket
+        # Listar todos los objetos en el bucket
         response = s3_client.list_objects_v2(Bucket=bucket_name)
         if 'Contents' in response:
+            deleted_count = 0
             for obj in response['Contents']:
-                s3_client.delete_object(Bucket=bucket_name, Key=obj['Key'])
-        print(f"Bucket '{bucket_name}' limpiado exitosamente.")
+                # Excluir objetos que estén en el directorio automorphisms/
+                if not obj['Key'].startswith('automorphisms/'):
+                    s3_client.delete_object(Bucket=bucket_name, Key=obj['Key'])
+                    deleted_count += 1
+            
+            print(f"Bucket '{bucket_name}' limpiado exitosamente.")
+            print(f"Se eliminaron {deleted_count} objetos (excluyendo directorio 'automorphisms/').")
+        else:
+            print(f"No hay objetos para eliminar en el bucket '{bucket_name}'.")
         
     except Exception as e:
         print(f"Error al limpiar el bucket '{bucket_name}': {e}")
